[dev] Gates; Level1; Crab enemy; SFX

+++ Assets/_Game/_Core/Configs/Scripts/CharacterSettingsSO.cs
@@ -5,7 +5,7 @@ namespace SoloGames.Configs
     [CreateAssetMenu(menuName = "Game/Configs/Character Settings", fileName = "Character Settings")]
     public class CharacterSettingsSO : ScriptableObject
     {
-        public int Health;
+        public float Health;
         public float Attack;
         public float MoveSpeed;
         public float JumpForce;
